refactor: use downloader
