+++ src/html.pre.js
@@ -12,6 +12,7 @@
 const fixSections = require('./fix-sections.js');
 const createPageBlocks = require('./create-page-blocks.js');
 const createPictures = require('./create-pictures.js');
+const fetchMetadata = require('./fetch-metadata.js');
 const extractMetaData = require('./extract-metadata.js');
 
 /**
@@ -39,4 +40,5 @@ async function pre(context, action) {
   await extractMetaData(context, action);
 }
 
+pre.before = { content: fetchMetadata };
 module.exports.pre = pre;
